% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeling_lucky.R
\name{feeling_lucky}
\alias{feeling_lucky}
\title{Print a random Sacred Harp song lyric}
\usage{
feeling_lucky()
}
\description{
Fetches a random lyric from the 2025 Sacred Harp edition on GitHub and prints it.
Called automatically when loading the database, or run it directly anytime.
}
