% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_minutes}
\alias{get_minutes}
\title{Query the fasola minutes database}
\usage{
get_minutes(
  query_fn = identity,
  include_location = FALSE,
  include_song_detail = FALSE,
  include_leader_detail = FALSE,
  name = "fasola_df"
)
}
\arguments{
\item{query_fn}{A function that takes a pre-joined tbl and returns a query.}

\item{include_location}{Add city, state_province, country, gps_lat, gps_long.}

\item{include_song_detail}{Add meter, music_attribution, three_liner.}

\item{include_leader_detail}{Add lesson_count, song_entropy, top20_count, location_count.}

\item{name}{Name of the variable assigned in the global environment.}
}
\description{
Query the fasola minutes database
}
